Update compileSdk, targetSdk and app_name
In this commit, I have updated the compileSdk and targetSdk from 33 to 34 in build.gradle.kts to align with the latest SDK version for better app performance and compatibility. I've also changed the 'app_name' in strings.xml from 'Template Android App' to 'QR Delivery' to correctly represent the purpose of the application.