Remove DetailItem and rename theme
DetailItem.kt file has been removed as it was not contributing to the desired application structure. The theme name in Theme.kt, HomeScreen.kt and MainActivity.kt was also refactored from TemplateAndroidAppTheme to QRDeliveryAppTheme to more accurately reflect the intended use of the application theme. AnticipateInterpolator was also removed from MainActivity.kt as it was unused.