Add authentication screen to the application
Implemented a new authentication screen to the application. The navigation route for the authentication screen has been added to the `MainNavGraph`. Screen consists of login options with Google and Apple, and appropriate drawables have been added. This update is essential to ensure user authentication before accessing the home screen. The `AuthScreen` compose utilizes reusable components for different login methods to improve code efficiency and design coherence.