Update navigation route from Home to Main
Modified the Home route to Main route in the MainNavGraph. This change is associated with introducing a BottomNav component to enhance app navigation. Now, on successful login, the navigation will redirect the user to the Main screen instead of the Home screen. This is done to provide users with a more integrated and user-friendly navigation system.