Add BottomNav component for easier navigation
A new file 'BottomNav.kt' was added as a component to the application. This new addition allows the navigation system to be more user-friendly and intuitive by introducing a navigation bar at the bottom of the screen. This will replace the previous system where the user was redirected straight to the Home screen. Now, the user will be redirected to the Main screen where they can easily navigate to other sections of the app using the BottomNav.