Add extension for NavGraphBuilder
Added a new file NavHost.kt which is an extension of the NavGraphBuilder. It creates a new composable function which enables a navigational content with animations. This change will make our code more reusable, organized and easier to read. It also enhances the navigation among different screens of the app.