Add ParcelPage.kt to implement parcel page view
Created a new Kotlin class 'ParcelPage.kt' in the 'qr_delivery' package. This class defines a composable function for the Parcel Page view in the app, which includes the basic layout placeholder with a text label. This change is part of the larger effort to modularize page-level components for better separation of concerns and increased readability.