Add primaryShadow extension to Modifier in UI
An extension function `primaryShadow` has been added to `Modifier` in the UI theme. The function is used to apply a shadow with primary color to a shape. The intent is to allow consistent shadow effects throughout the app. Adding this extension aids in maintaining a coherent aesthetic throughout the application and simplifies the modification of the shadows' properties.