Add Firebase services to project
Added Firebase Analytics, Crashlytics, and Performance to the project. This was achieved by adding relevant dependencies to the gradle files, adding the version of the firebase-bom to the gradle.properties file, and enabling logcat in the Android manifest file. These additions will allow for better tracking, crash reporting, and performance monitoring of the application.