Added CameraX dependencies and permissions
CameraX dependencies have been added to the build.gradle script to include camera functionality within the application. Defined `cameraXVersion` in gradle.properties file to ensure a consistent CameraX version across all modules. Added necessary permissions and uses-sdk entries in the AndroidManifest.xml file required for the use of the camera in the app.