Add new "compose-permissions" library module
This commit adds a new "compose-permissions" library module to the application. This module is meant to handle all permissions related functionality for the entire application. The build.gradle file for the module includes necessary dependencies, compiles SDK steps and minimum SDK version, along with test implementation details. The library plugin for android and kotlin has been added to build.gradle.kts. An empty AndroidManifest.xml file is also included as an initial setup. The module was included in the settings.gradle file. Also, the initial template for unit and instrumental tests has been setup inside the module.