Adjust order of plugin declarations in Gradle file
Reordered the plugin declarations in our Gradle file to align with the order they are used in our project for better readability. The 'com.android.library' plugin was not at an intuitive place in our code, making it harder to find and modify settings related to it. This new arrangement improves the ease of managing these configurations.